test/pi-zero/requirements.txt
paho-mqtt==1.5.1
bleak==0.12.1
ntplib==0.4.0
